Libft

+++ Libft/ft_strlen.c
@@ -10,9 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
-size_t	ft_strlen(const char *str)
+unsigned int	ft_strlen(const char *str)
 {
 	int	i;
