+++ Libft/ft_substr.c
@@ -10,4 +10,4 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-char *ft_substr(char const *s, unsigned int start, size_t len);
+char *ft_substr(char const *s, unsigned int start, unsigned int len);
